feat: complete setting product

+++ controllers/SettingProductController.js
@@ -1,4 +1,6 @@
 const SettingProduct = require('../models/SettingProduct')
+const SettingProductCategory = require('../models/SettingProductCategory')
+const { checkExistBetweenTwoArray } = require('../utils/helpers')
 
 module.exports.getAll = async (req, res, next) => {
   try {
@@ -10,10 +12,20 @@ module.exports.getAll = async (req, res, next) => {
 }
 
 module.exports.create = async (req, res, next) => {
+  const { value, categories } = req.body
   try {
-    const productSettingExist = await SettingProduct.findOne({ value: req.body.value })
+    const productSettingExist = await SettingProduct.findOne({ value })
     if (productSettingExist) {
       return res.status(404).json({ status: 'error', message: 'Value Setting Product Exist' })
+    }
+
+    const getAllCategories = await SettingProductCategory.find().select('_id')
+    const newMap = getAllCategories.map((item) => item._id.toString())
+
+    if (!checkExistBetweenTwoArray(newMap, categories)) {
+      return res
+        .status(404)
+        .json({ status: 'error', message: 'Setting Product Category Not Found' })
     }
 
     const productSetting = await SettingProduct.create(req.body)
@@ -29,18 +41,33 @@ module.exports.create = async (req, res, next) => {
 
 module.exports.updateSetting = async (req, res, next) => {
   const { settingProductId } = req.params
-  const { value, categories } = req.body
+  const { value, categories, priceIncrease } = req.body
   try {
     const settingProduct = await SettingProduct.findOne({ _id: settingProductId })
-    settingProduct.value = value
-    settingProduct.categories = categories
+    if (!settingProduct) {
+      return res.status(200).json({
+        status: 'error',
+        message: 'Setting Product Not Found',
+      })
+    }
 
-    const afterSetting = await settingProduct.save()
-    return res.status(200).json({
-      status: 'success',
-      data: afterSetting,
-      message: 'Update Setting Product',
-    })
+    const getAllCategories = await SettingProductCategory.find().select('_id')
+    const newMap = getAllCategories.map((item) => item._id.toString())
+
+    if (!checkExistBetweenTwoArray(newMap, categories)) {
+      return res
+        .status(404)
+        .json({ status: 'error', message: 'Setting Product Category Not Found' })
+    } else {
+      settingProduct.value = value
+      settingProduct.categories = categories
+      settingProduct.priceIncrease = priceIncrease
+      await settingProduct.save()
+      return res.status(200).json({
+        status: 'success',
+        message: 'Update Setting Product',
+      })
+    }
   } catch (error) {
     return res.status(500).json({ status: 'error', message: error.message })
   }
@@ -49,10 +76,17 @@ module.exports.updateSetting = async (req, res, next) => {
 module.exports.deleteSetting = async (req, res, next) => {
   const { settingProductId } = req.params
   try {
+    const settingProduct = await SettingProduct.findOne({ _id: settingProductId })
+    if (!settingProduct) {
+      return res.status(200).json({
+        status: 'error',
+        message: 'Setting Product Not Found',
+      })
+    }
+
     await SettingProduct.deleteOne({ _id: settingProductId })
     return res.status(200).json({
       status: 'success',
-
       message: 'Deleted Setting Product',
     })
   } catch (error) {
